task stdout and stderr are now written to (mostly-compatible) GridFS documents.
These can be read by mongofiles (v3.0.7 tested) and the soon-deprecated MongoDB::GridFS, but the in-development MongoDB::GridFSBucket (v1.3.1) breaks on them.

The collection prefix for this is 'task_output'.

The chunks are stored as strings. As such if non-UTF8 content is in them, that part of the content will not be saved correctly and not read properly outside of Perl. But in general that should not occur, and this allows regex searches on the chunk contents.

This can be disabled by setting 'gfs: false' in the config.